cambio double por bigdecimal

+++ ms-facturacion/src/main/java/edu/tudai/microserviciofacturacion/DTO/DetalleFacturaDTO.java
@@ -2,13 +2,15 @@ package edu.tudai.microserviciofacturacion.DTO;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 public class DetalleFacturaDTO {
 
     private Long facturaId;
     private Long viajeId;
-    private Double tarifaBase;
-    private Double tarifaExtra;
+    private BigDecimal tarifaBase;
+    private BigDecimal tarifaExtra;
     private Long tiempoUso;
     private Long tiempoPausado;
